Rename github avatar to import
