+++ src/pages/branch/CreateBranch.tsx
@@ -12,7 +12,7 @@ import {
 } from "firebase/firestore";
 import firebaseapi from "../../utils/firebaseapi";
 import TiptapEditor from "../../components/editor/Editor";
-import defaultAvatar from "../../assets/images/defaultAvatar.png";
+import defaultAvatar from "../../assets/images/githubAvatar.png";
 import {
   Button,
   MergeBtn,
